+++ CHANGELOG.md
@@ -9,6 +9,9 @@ https://github.com/The-Network-Crew/Proxmox-VE-for-WHMCS/milestones
 - Max Memory: Ensure you can set more than 128GB (#167)
 - Max CPUs/Cores: Change columns to 3 characters (#167)
 
+### 🐛 Bug Fix
+- noVNC: Delete PVEAuthCookie before setting it (#167)
+
 ## [1.2.16] - 2025-10-15 - _"Minor Adjustments"_
 
 ### 💅 Polish